Use as evm's origin on system evm calsl
